requirements.txt
numpy~=1.22.3           	  # Basic mathematical library for array operations
scipy~=1.8.0            	  # Scientific library
matplotlib~=3.5.2       	  # Plotting library
h5py~=3.6.0             	  # Data storage format
pandas~=1.4.2           	  # Dataframe library
python-dotenv~=0.20.0   	  # Configuration library
hydra-core~=1.2.0       	  # Configuration library
hydra-joblib-launcher~=1.2.0      # parallel jobs locally
hydra-submitit-launcher~=1.1.6    # parallel jobs on HPC cluster
pytorch-lightning~=1.6.3          # only for LightningDataModule
omegaconf~=2.2.1       	 	  # Configuration library
clawpack~=5.6.1         	  # Finite Volume library
deepxde~=1.1.3          	  # Physics-informed ML library
torch~=1.11.0           	  # PyTorch library for ML applications
jax~=0.3.13             	  # Numerical library
jaxlib~=0.3.10          	  # Numerical library part 2 (CPU)
dash~=2.2.0             	  # dashboard visualisation for phiflow
phiflow~=2.0.3          	  # PDE simulator
pyDaRUS~=1.0.5          	  # DaRUS API package to upload and download dataset
pyDataverse@ git+https://github.com/JR-1991/pyDataverse.git@0fcfcd3fbc6bf1aec869899f715a51dca25e91be
imageio~=2.19.2
tqdm~=4.64.0
einops~=0.4.1
tensorboard~=2.9.0
tensorboard-plugin-profile~=2.8.0

# Optional, if you want to use jupyter
jupyter~=1.0.0
ipykernel~=6.9.0
jupyter-dash~=0.4.1
nbstripout~=0.5.0       # trim outputs from jupyter notebooks